Stop LLM from loading automatically on server start
- Change initialize_lifecycle() default preload_llm from True to False

- Remove duplicate preload_llm_model() call from daemon.py

- Extract ProcessManagerMixin and KnowledgeMigrationMixin from runtime_mixin.py to fix pre-existing code size violations
